i18n: make catalog compile work from setup.py
