+++ setup.cfg
@@ -1,2 +1,10 @@
 [install]
 optimize = 1
+
+[compile_catalog]
+directory = indico/translations
+
+[compile_catalog_js]
+input_dir = indico/translations
+output_dir = indico/htdocs/js/indico/i18n
+domain = messages-js
